requirements.txt
flask
pymongo
python-dotenv
requests
pandas
fastf1